Move Codable conformance out of PackageModel+Codable.swift
